# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Anime Pipeline API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: episodes
    description: Episode management
  - name: pipeline
    description: Pipeline control
  - name: sites
    description: Anime site configuration
  - name: logs
    description: Pipeline logs
  - name: runs
    description: Pipeline run history
  - name: scheduler
    description: Scheduler configuration
  - name: whitelist
    description: Anime whitelist management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /episodes:
    get:
      operationId: listEpisodes
      tags: [episodes]
      summary: List all episodes
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
            enum: [pending, downloading, downloaded, sending, sent, failed, null]
        - name: animeName
          in: query
          schema:
            type: ["string", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of episodes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EpisodeListResponse"

  /episodes/export:
    get:
      operationId: exportEpisodes
      tags: [episodes]
      summary: Export episodes as JSON for CSV download
      responses:
        "200":
          description: All episodes for export
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Episode"

  /episodes/{id}:
    get:
      operationId: getEpisode
      tags: [episodes]
      summary: Get episode by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Episode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEpisode
      tags: [episodes]
      summary: Delete episode record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /episodes/{id}/retry:
    post:
      operationId: retryEpisode
      tags: [episodes]
      summary: Retry failed episode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Episode"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pipeline/run:
    post:
      operationId: runPipeline
      tags: [pipeline]
      summary: Trigger full pipeline run
      responses:
        "200":
          description: Pipeline started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatus"

  /pipeline/download:
    post:
      operationId: triggerDownload
      tags: [pipeline]
      summary: Manually trigger download for a specific episode URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualDownloadRequest"
      responses:
        "200":
          description: Download started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatus"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /pipeline/status:
    get:
      operationId: getPipelineStatus
      tags: [pipeline]
      summary: Get current pipeline status
      responses:
        "200":
          description: Pipeline status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStatus"

  /pipeline/stats:
    get:
      operationId: getPipelineStats
      tags: [pipeline]
      summary: Get pipeline statistics
      responses:
        "200":
          description: Statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PipelineStats"

  /pipeline/activity:
    get:
      operationId: getPipelineActivity
      tags: [pipeline]
      summary: Get daily activity data for charts
      parameters:
        - name: days
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Activity data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityPoint"

  /sites:
    get:
      operationId: listSites
      tags: [sites]
      summary: List configured anime sites
      responses:
        "200":
          description: List of sites
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnimeSite"
    post:
      operationId: createSite
      tags: [sites]
      summary: Add a new anime site
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAnimeSiteRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeSite"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /sites/{id}:
    patch:
      operationId: updateSite
      tags: [sites]
      summary: Update site configuration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAnimeSiteRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnimeSite"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSite
      tags: [sites]
      summary: Remove a site
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /logs:
    get:
      operationId: listLogs
      tags: [logs]
      summary: Get pipeline logs
      parameters:
        - name: level
          in: query
          schema:
            type: ["string", "null"]
            enum: [info, warning, error, success, null]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineLog"

  /runs:
    get:
      operationId: listRuns
      tags: [runs]
      summary: List pipeline run history
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Run history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PipelineRun"

  /scheduler:
    get:
      operationId: getScheduler
      tags: [scheduler]
      summary: Get scheduler configuration
      responses:
        "200":
          description: Scheduler config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerConfig"
    patch:
      operationId: updateScheduler
      tags: [scheduler]
      summary: Update scheduler configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSchedulerRequest"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchedulerConfig"

  /whitelist:
    get:
      operationId: listWhitelist
      tags: [whitelist]
      summary: List anime whitelist
      responses:
        "200":
          description: Whitelist entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WhitelistEntry"
    post:
      operationId: createWhitelistEntry
      tags: [whitelist]
      summary: Add anime to whitelist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWhitelistRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /whitelist/{id}:
    patch:
      operationId: updateWhitelistEntry
      tags: [whitelist]
      summary: Update whitelist entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWhitelistRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WhitelistEntry"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWhitelistEntry
      tags: [whitelist]
      summary: Remove from whitelist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    Episode:
      type: object
      properties:
        id:
          type: integer
        animeName:
          type: string
        season:
          type: integer
        episode:
          type: integer
        sourceUrl:
          type: string
        videoUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, downloading, downloaded, sending, sent, failed]
        errorMessage:
          type: ["string", "null"]
        filePath:
          type: ["string", "null"]
        telegramMessageId:
          type: ["string", "null"]
        quality:
          type: ["string", "null"]
        priority:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, animeName, season, episode, sourceUrl, status, priority, createdAt, updatedAt]

    EpisodeListResponse:
      type: object
      properties:
        episodes:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
        total:
          type: integer
      required: [episodes, total]

    PipelineStatus:
      type: object
      properties:
        running:
          type: boolean
        message:
          type: string
        startedAt:
          type: ["string", "null"]
          format: date-time
      required: [running, message]

    PipelineStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        downloading:
          type: integer
        downloaded:
          type: integer
        sending:
          type: integer
        sent:
          type: integer
        failed:
          type: integer
        totalAnimes:
          type: integer
        recentActivity:
          type: array
          items:
            $ref: "#/components/schemas/Episode"
      required: [total, pending, downloading, downloaded, sending, sent, failed, totalAnimes, recentActivity]

    ActivityPoint:
      type: object
      properties:
        date:
          type: string
        downloaded:
          type: integer
        sent:
          type: integer
        failed:
          type: integer
      required: [date, downloaded, sent, failed]

    ManualDownloadRequest:
      type: object
      properties:
        animeName:
          type: string
        season:
          type: integer
        episode:
          type: integer
        sourceUrl:
          type: string
        quality:
          type: ["string", "null"]
        priority:
          type: ["integer", "null"]
      required: [animeName, season, episode, sourceUrl]

    AnimeSite:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        baseUrl:
          type: string
        enabled:
          type: boolean
        requiresJs:
          type: boolean
        scraperType:
          type: string
        lastScrapedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, baseUrl, enabled, requiresJs, scraperType, createdAt]

    CreateAnimeSiteRequest:
      type: object
      properties:
        name:
          type: string
        baseUrl:
          type: string
        enabled:
          type: boolean
        requiresJs:
          type: boolean
        scraperType:
          type: string
      required: [name, baseUrl, scraperType]

    UpdateAnimeSiteRequest:
      type: object
      properties:
        name:
          type: ["string", "null"]
        baseUrl:
          type: ["string", "null"]
        enabled:
          type: ["boolean", "null"]
        requiresJs:
          type: ["boolean", "null"]
        scraperType:
          type: ["string", "null"]

    PipelineLog:
      type: object
      properties:
        id:
          type: integer
        level:
          type: string
          enum: [info, warning, error, success]
        message:
          type: string
        details:
          type: ["string", "null"]
        episodeId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, level, message, createdAt]

    PipelineRun:
      type: object
      properties:
        id:
          type: integer
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: ["string", "null"]
          format: date-time
        episodesFound:
          type: integer
        episodesDownloaded:
          type: integer
        episodesFailed:
          type: integer
        status:
          type: string
          enum: [running, completed, failed]
        trigger:
          type: string
          enum: [manual, scheduler, api]
        durationSeconds:
          type: ["integer", "null"]
      required: [id, startedAt, episodesFound, episodesDownloaded, episodesFailed, status, trigger]

    SchedulerConfig:
      type: object
      properties:
        id:
          type: integer
        enabled:
          type: boolean
        intervalMinutes:
          type: integer
        lastRunAt:
          type: ["string", "null"]
          format: date-time
        nextRunAt:
          type: ["string", "null"]
          format: date-time
        discordWebhookUrl:
          type: ["string", "null"]
        notifyOnError:
          type: boolean
        defaultQuality:
          type: string
          enum: [best, 1080p, 720p, 480p, worst]
        useWhitelist:
          type: boolean
      required: [id, enabled, intervalMinutes, notifyOnError, defaultQuality, useWhitelist]

    UpdateSchedulerRequest:
      type: object
      properties:
        enabled:
          type: ["boolean", "null"]
        intervalMinutes:
          type: ["integer", "null"]
        discordWebhookUrl:
          type: ["string", "null"]
        notifyOnError:
          type: ["boolean", "null"]
        defaultQuality:
          type: ["string", "null"]
          enum: [best, 1080p, 720p, 480p, worst, null]
        useWhitelist:
          type: ["boolean", "null"]

    WhitelistEntry:
      type: object
      properties:
        id:
          type: integer
        animeName:
          type: string
        priority:
          type: integer
        qualityPreference:
          type: string
          enum: [best, 1080p, 720p, 480p, worst]
        enabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, animeName, priority, qualityPreference, enabled, createdAt]

    CreateWhitelistRequest:
      type: object
      properties:
        animeName:
          type: string
        priority:
          type: ["integer", "null"]
        qualityPreference:
          type: ["string", "null"]
          enum: [best, 1080p, 720p, 480p, worst, null]
        enabled:
          type: ["boolean", "null"]
      required: [animeName]

    UpdateWhitelistRequest:
      type: object
      properties:
        animeName:
          type: ["string", "null"]
        priority:
          type: ["integer", "null"]
        qualityPreference:
          type: ["string", "null"]
          enum: [best, 1080p, 720p, 480p, worst, null]
        enabled:
          type: ["boolean", "null"]