# lib/api-zod/src/generated/types/createAnimeSiteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */

export interface CreateAnimeSiteRequest {
  name: string;
  baseUrl: string;
  enabled?: boolean;
  requiresJs?: boolean;
  scraperType: string;
}