# lib/api-zod/src/generated/types/createWhitelistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */
import type { CreateWhitelistRequestQualityPreference } from "./createWhitelistRequestQualityPreference";

export interface CreateWhitelistRequest {
  animeName: string;
  /** @nullable */
  priority?: number | null;
  /** @nullable */
  qualityPreference?: CreateWhitelistRequestQualityPreference;
  /** @nullable */
  enabled?: boolean | null;
}