# lib/api-zod/src/generated/types/manualDownloadRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */

export interface ManualDownloadRequest {
  animeName: string;
  season: number;
  episode: number;
  sourceUrl: string;
  /** @nullable */
  quality?: string | null;
  /** @nullable */
  priority?: number | null;
}