# lib/api-zod/src/generated/types/pipelineRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */
import type { PipelineRunStatus } from "./pipelineRunStatus";
import type { PipelineRunTrigger } from "./pipelineRunTrigger";

export interface PipelineRun {
  id: number;
  startedAt: Date;
  /** @nullable */
  endedAt?: Date | null;
  episodesFound: number;
  episodesDownloaded: number;
  episodesFailed: number;
  status: PipelineRunStatus;
  trigger: PipelineRunTrigger;
  /** @nullable */
  durationSeconds?: number | null;
}