# lib/api-zod/src/generated/types/pipelineStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */
import type { Episode } from "./episode";

export interface PipelineStats {
  total: number;
  pending: number;
  downloading: number;
  downloaded: number;
  sending: number;
  sent: number;
  failed: number;
  totalAnimes: number;
  recentActivity: Episode[];
}