# lib/api-zod/src/generated/types/schedulerConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */
import type { SchedulerConfigDefaultQuality } from "./schedulerConfigDefaultQuality";

export interface SchedulerConfig {
  id: number;
  enabled: boolean;
  intervalMinutes: number;
  /** @nullable */
  lastRunAt?: Date | null;
  /** @nullable */
  nextRunAt?: Date | null;
  /** @nullable */
  discordWebhookUrl?: string | null;
  notifyOnError: boolean;
  defaultQuality: SchedulerConfigDefaultQuality;
  useWhitelist: boolean;
}