# lib/api-zod/src/generated/types/updateAnimeSiteRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Anime Pipeline API
 * OpenAPI spec version: 0.2.0
 */

export interface UpdateAnimeSiteRequest {
  /** @nullable */
  name?: string | null;
  /** @nullable */
  baseUrl?: string | null;
  /** @nullable */
  enabled?: boolean | null;
  /** @nullable */
  requiresJs?: boolean | null;
  /** @nullable */
  scraperType?: string | null;
}